Add Missing Files and Action Return Values
Performing an action will not only return the next action, but they
will also return a flag indicating whether the screen is dirty. This
flag will be enabled when an animation is currently in progress.
It will be false when no animation is going on and thus the program
can save cycles by waiting for events instead.