Refactoring of Chart and Header Functions
Clearly define inputs and outputs when drawing the chart and its
header. Previously, the functions took a state object and returned
the next state object. Some fields of the state record were being
used as inputs but weren't needed to maintain state. Separate
inputs from state and return an output with the next state instead.