Use FTGL without Layouts
The FTGL binding does not have the getLayoutBBox binding, because the
C function is misspelled in the stable release of FTGL. Furthermore,
there is a bug where the bounding box does not match the text when
rendered. As a result, avoid using layouts and rely on simpler
resizing logic where the content just overflows rather than trying
to rearrange itself. This turns out to make the code a lot simpler.