New Task Management Algorithm
Do the tasks sequentially in the background. It was the parsing of
the CSV data that was slowing the animation down. In the future,
add a way to distinguish between heavy and light tasks, where
light tasks could be bundled together in a single batch while
heavy tasks will require one thread alone.